requirements.txt
flask==3.0.2
gunicorn==21.2.0
numpy==1.24.3
opencv-python-headless==4.9.0.80
pillow==10.2.0
tensorflow==2.15.0
